Dang Hoang version 2 28.12.2022

+++ src/main/java/com/java04/wibucian/services/EmployeeService.java
@@ -51,8 +51,15 @@ public class EmployeeService {
         return this.employeeRepository.findAll();
     }
 
-    public void delete(String id) {
+    public boolean delete(String id) {
+        Employee employee = employeeRepository.findById(id).orElse(null);
+        if (employee == null) {
+            return false;
+        }
+        String idAccount = employee.getAccount().getId();
+        accountRepository.deleteById(idAccount);
         employeeRepository.deleteById(id);
+        return true;
     }
 
     public void update(String id, EmployeeUpdateVO vO) {
